test(core-api): test plugins
