+++ __tests__/unit/core-api/__support__/server.ts
@@ -2,7 +2,7 @@ import { Server } from "@packages/core-api/src";
 import { preparePlugins } from "@packages/core-api/src/plugins";
 import { Application, Container, Providers } from "@packages/core-kernel";
 
-export const initServer = async (app: Application, defaults: any, customRoute: any): Promise<Server> => {
+export const initServer = async (app: Application, defaults: any, customRoute?: any): Promise<Server> => {
     const serverConfig = {
         enabled: true,
         host: "0.0.0.0",
@@ -14,6 +14,8 @@ export const initServer = async (app: Application, defaults: any, customRoute: a
 
     app.bind(Server).to(Server);
     const server = app.get(Server);
+    // @ts-ignore
+    server.app.app = app;
 
     const pluginConfiguration = app.get<Providers.PluginConfiguration>(Container.Identifiers.PluginConfiguration);
     const pluginConfigurationInstance: Providers.PluginConfiguration = pluginConfiguration.from("core-api", defaults);
